Refactor into individual classes
